+++ backend/pages/deadline.py
@@ -1,6 +1,7 @@
 from flask import Blueprint, jsonify, request
 from modules.user import encode_jwt, decode_jwt
 import requests
+from app.models import deadlines
 from app import db
 from sqlalchemy.sql import text
 import json
@@ -29,16 +30,3 @@ def get_deadlines(username):
     for i in range(len(deadlines)):
         response[i] = {"id": deadlines[i][0], "username": deadlines[i][1], "course_id": deadlines[i][2], "date": deadlines[i][3]}
     return json.dumps(response, default=str)
-
-# Database model demo
-class deadlines(db.Model):
-    __tablename__ = 'deadlines'
-    __table_args__ = {'extend_existing': True} 
-    id = db.Column(db.Integer, primary_key=True)
-    username = db.Column(db.String, nullable=False)
-    course_id = db.Column(db.Integer, nullable=False)
-    date = db.Column(db.DateTime, nullable=False)
-
-
-    def __repr__(self):
-        return f"deadlines('{self.date}')"
