+++ backend/pages/main.py
@@ -21,15 +21,3 @@ def dbdemo():
     db.session.add(target)
     db.session.commit()
     return jsonify(message="Database demo successful!")
-
-
-# Database model demo
-class TargetDate(db.Model):
-    __tablename__ = "target_date"
-    id = db.Column(db.Integer, primary_key=True)
-    username = db.Column(db.String, nullable=False)
-    course = db.Column(db.String, nullable=False)
-    date = db.Column(db.DateTime, nullable=False)
-
-    def __repr__(self):
-        return f"target_date('{self.date}')"
